backend/requirements.txt
flask==2.0.3
flask-pymongo==2.3.0
flask-cors==3.0.10
python-dotenv==0.19.0
werkzeug==2.0.3
pyjwt==2.3.0 